fix(LocationContext): move to correct folder, since it's a hook
